requirements.txt
streamlit
pandas
plotly
pywinrm
requests
google-generativeai